+++ MoreFertilizers/MoreFertilizers/docs/Changelog.md
@@ -12,6 +12,7 @@
 #### Version 0.3.3
 * Update to latest AtraCore.
 * Fix everlasting fertilizer not getting a special color.
+* Fix Radioactive fertilizer not properly ignoring seasons in Greenhouse locations.
 
 #### Version 0.3.2
 * Balance